+++ src/api/server.js
@@ -9,7 +9,7 @@ const groupAPI = require('./group');
 const tagAPI = require('./tag');
 const notificationAPI = require('./notification');
 const templateAPI = require('./template');
-const webhookAPI = require('./_webhook');
+const webhookAPI = require('./webhook');
 const koaJsend = require('./middleware/jsend');
 
 const app = new Koa();
@@ -42,9 +42,9 @@ router.get('template/:id', templateAPI.get);
 router.put('template/:id', templateAPI.update);
 router.delete('template/:id', templateAPI.delete);
 
-router.post('_webhook', webhookAPI.createSchema, webhookAPI.create);
-router.get('_webhook', webhookAPI.get);
-router.delete('_webhook/:url', webhookAPI.delete);
+router.post('webhook', webhookAPI.createSchema, webhookAPI.create);
+router.get('webhook', webhookAPI.get);
+router.delete('webhook/:url', webhookAPI.delete);
 
 router.get('ping', async ctx => ctx.success({ time: Math.floor(Date.now() / 1000) }));
